Add files via upload
I started this app again and hopefully for the last time. The difference here and from now on is that this is all my very own code - the way I see and understand things. As a personal note to self - this is the moment you let go of the safety wheels, no more copy pasting. 
At this point, all we have is a function that vardumps an item (in  $_SESSION) as an array that has the item and the due date if the add task button is pressed. also note that it took 24 minutes from start to finishing with html/bootstrap and 1hr 30min from then on to now (with little breaks) I'm not sure where to from here but it's 2am and I'd rather carry on than sleep. who am I?