Add files via upload
In 22 minutes I have: an add to do item button and function that adds a NEW array of an item with due date and item. wild. I am busy with foreach output onto list. currently working but not satisfactory, getting a warning so saving work before I break things. also computer is complaining so don't want to take chances.